+++ custom_components/sber_mqtt_bridge/www/components/sber-detail-dialog.js
@@ -5,11 +5,7 @@
  * device registry data, model config with features and allowed values.
  */
 
-const LitElement = Object.getPrototypeOf(
-  customElements.get("ha-panel-lovelace") ?? customElements.get("hui-view")
-);
-const html = LitElement?.prototype.html;
-const css = LitElement?.prototype.css;
+import { LitElement, html, css } from "../lit-base.js";
 
 class SberDetailDialog extends LitElement {
   static get properties() {