+++ custom_components/sber_mqtt_bridge/www/components/sber-diagnose.js
@@ -12,11 +12,7 @@
  * entity.
  */
 
-const LitElement = Object.getPrototypeOf(
-  customElements.get("ha-panel-lovelace") ?? customElements.get("hui-view")
-);
-const html = LitElement?.prototype.html;
-const css = LitElement?.prototype.css;
+import { LitElement, html, css } from "../lit-base.js";
 
 const VERDICT_LABEL = {
   ok: "Clean",