+++ custom_components/sber_mqtt_bridge/www/components/sber-link-dialog.js
@@ -6,11 +6,7 @@
  * Fires "links-saved" event when links are updated.
  */
 
-const LitElement = Object.getPrototypeOf(
-  customElements.get("ha-panel-lovelace") ?? customElements.get("hui-view")
-);
-const html = LitElement?.prototype.html;
-const css = LitElement?.prototype.css;
+import { LitElement, html, css } from "../lit-base.js";
 
 class SberLinkDialog extends LitElement {
   static get properties() {