+++ custom_components/sber_mqtt_bridge/www/components/sber-validation.js
@@ -17,11 +17,7 @@
  *                 happened during this debug session".
  */
 
-const LitElement = Object.getPrototypeOf(
-  customElements.get("ha-panel-lovelace") ?? customElements.get("hui-view")
-);
-const html = LitElement?.prototype.html;
-const css = LitElement?.prototype.css;
+import { LitElement, html, css } from "../lit-base.js";
 
 class SberValidation extends LitElement {
   static get properties() {